+++ packages/db/migrations/0003_broken_grim_reaper.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "crypto_config" ADD COLUMN "auto_generate" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "crypto_config" ADD COLUMN "auto_tick" boolean DEFAULT false NOT NULL;